+++ src/Class/SueloMuestra.java
@@ -12,16 +12,16 @@ package Class;
 public class SueloMuestra {
 
     // atributos Volumen
-    private int Vol_Vacios;
-    private int Vol_Agua;
-    private int Vol_Solidos;
-    private int Vol_Aire;
-    private int Vol_Total;
+    private double Vol_Vacios;
+    private double Vol_Agua;
+    private double Vol_Solidos;
+    private double Vol_Aire;
+    private double Vol_Total;
     //atributos Peso
-    private int Pes_Agua;
-    private int Pes_Solidos;
-    private int Pes_Aire;
-    private int Pes_Total;
+    private double Pes_Agua;
+    private double Pes_Solidos;
+    private double Pes_Aire;
+    private double Pes_Total;
 
     //Constructor
     public SueloMuestra() {
@@ -37,43 +37,43 @@ public class SueloMuestra {
     }
 
     //Metodos Volumen
-    public int getVol_Vacios() {
+    public double getVol_Vacios() {
         return Vol_Vacios;
     }
 
-    public void setVol_Vacios(int Vol_Vacios) {
+    public void setVol_Vacios(double Vol_Vacios) {
         this.Vol_Vacios = Vol_Vacios;
     }
 
-    public int getVol_Agua() {
+    public double getVol_Agua() {
         return Vol_Agua;
     }
 
-    public void setVol_Agua(int Vol_Agua) {
+    public void setVol_Agua(double Vol_Agua) {
         this.Vol_Agua = Vol_Agua;
     }
 
-    public int getVol_Solidos() {
+    public double getVol_Solidos() {
         return Vol_Solidos;
     }
 
-    public void setVol_Solidos(int Vol_Solidos) {
+    public void setVol_Solidos(double Vol_Solidos) {
         this.Vol_Solidos = Vol_Solidos;
     }
 
-    public int getVol_Aire() {
+    public double getVol_Aire() {
         return Vol_Aire;
     }
 
-    public void setVol_Aire(int Vol_Aire) {
+    public void setVol_Aire(double Vol_Aire) {
         this.Vol_Aire = Vol_Aire;
     }
 
-    public int getVol_Total() {
+    public double getVol_Total() {
         return Vol_Total;
     }
 
-    public void setVol_Total(int Vol_Total) {
+    public void setVol_Total(double Vol_Total) {
         if (getVol_Aire() != 0 && getVol_Agua() != 0 && getVol_Solidos() != 0) {
             this.Vol_Total = getVol_Aire() + getVol_Agua() + getVol_Solidos();
         } else if (getVol_Vacios() != 0 && getVol_Solidos() != 0) {
@@ -84,35 +84,35 @@ public class SueloMuestra {
     }
     //Metodos Peso
 
-    public int getPes_Agua() {
+    public double getPes_Agua() {
         return Pes_Agua;
     }
 
-    public void setPes_Agua(int Pes_Agua) {
+    public void setPes_Agua(double Pes_Agua) {
         this.Pes_Agua = Pes_Agua;
     }
 
-    public int getPes_Solidos() {
+    public double getPes_Solidos() {
         return Pes_Solidos;
     }
 
-    public void setPes_Solidos(int Pes_Solidos) {
+    public void setPes_Solidos(double Pes_Solidos) {
         this.Pes_Solidos = Pes_Solidos;
     }
 
-    public int getPes_Aire() {
+    public double getPes_Aire() {
         return Pes_Aire;
     }
 
-    public void setPes_Aire(int Pes_Aire) {
+    public void setPes_Aire(double Pes_Aire) {
         this.Pes_Aire = Pes_Aire;
     }
 
-    public int getPes_Total() {
+    public double getPes_Total() {
         return Pes_Total;
     }
 
-    public void setPes_Total(int Pes_Total) {
+    public void setPes_Total(double Pes_Total) {
         this.Pes_Total = Pes_Total;
     }
